CR: Place tests and code being tested in the same directory for ease of understanding, add a README
