+++ sdk/python/lib/test/langhost/empty/__init__.py
@@ -11,20 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from os import path
-from .util import LanghostTest
-
-
-class OneResourceTest(LanghostTest):
-    def test_one_resource(self):
-        self.run_test(
-            program=path.join(self.base_path(), "one_resource"),
-            expected_resource_count=1)
-
-    def register_resource(self, _ctx, _dry_run, ty, name, _resource,
-                          _dependencies):
-        self.assertEqual(ty, "test:index:MyResource")
-        self.assertEqual(name, "testResource1")
-        return {
-            "urn": self.make_urn(ty, name),
-        }
