+++ sdk/python/lib/test/langhost/empty/test_empty.py
@@ -1,0 +1,28 @@
+# Copyright 2016-2018, Pulumi Corporation.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from os import path
+from ..util import LanghostTest
+
+
+class EmptyTests(LanghostTest):
+    def test_empty(self):
+        self.run_test(
+            program=path.join(self.base_path(), "empty"),
+            expected_resource_count=0)
+
+    def test_empty_pwd(self):
+        self.run_test(
+            pwd=path.join(self.base_path(), "empty"),
+            expected_resource_count=0,
+            program="./")
