+++ sdk/python/lib/test/langhost/one_complex_resource/__main__.py
