+++ sdk/python/lib/test/langhost/one_resource/__main__.py
