+++ sdk/python/lib/test/langhost/resource_thens/test_resource_thens.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 from os import path
 import unittest
-from .util import LanghostTest
+from ..util import LanghostTest
 
 from pulumi.runtime.rpc import Unknown
