+++ sdk/python/lib/test/langhost/ten_resources/__main__.py
