+++ sdk/python/lib/test/langhost/ten_resources/test_ten_resources.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from os import path
-from .util import LanghostTest
+from ..util import LanghostTest
 
 
 class TenResourcesTest(LanghostTest):
